fix(workspaces): reorder workspaces after rename
Fix a bug when reorder workspaces on Niri with the overview. If a
windows is moved creating a new workspace before a exiting one, the
client would create a new workspaces after and with the same name as the
existing one, and then rename the exiting workspace. Because the rename
wasn't triggering a reorder, the workspaces would end up in the wrong
order.